perf(utils): improve performance of videoSrc and videoCaption formatting
Replaces slower implementation of videoSrc and videoCaption formatting utils with simpler, more
efficient implementations. The primary goal was 1: to replace the usage of Array.reduce with a
simple for loop, and 2: to replace the implementation of casting values to an array via [].concat()
with a simple check where we only create a new array if Array.isArray is false. These are
micro-optimizations that won't make a noticeable difference to the end user but give me better
peace of mind.