fix(hovervideoplayer): remove src cleanup logic when component unmounts
Removes logic that would clear all of the video's sources and load to force the video to purge any
potential memory leaks. Overall, the problem being solved here isn't super major and it was causing
errors to be thrown as seen in #16 where an error would occur if video.load() was called while a
play promise was still in progress. If the removal of this cleanup code causes notable performance
issues we can revisit, but for now this should work just fine.

fixes #16